+++ core/src/elements/ons-alert-dialog/index.js
@@ -289,8 +289,8 @@ class AlertDialogElement extends BaseElement {
    * @property disabled
    * @type {Boolean}
    * @description
-   *   [en]A boolean value that specifies whether the dialog is disabled or not.[/en]
-   *   [ja][/ja]
+   *   [en]Whether the element is disabled or not.[/en]
+   *   [ja]無効化されている場合に`true`。[/ja]
    */
   set disabled(value) {
     return util.toggleAttribute(this, 'disabled', value);
@@ -304,11 +304,7 @@ class AlertDialogElement extends BaseElement {
    * @property cancelable
    * @type {Boolean}
    * @description
-   *   [en]
-   *     A boolean value that specifies whether the dialog is cancelable or not.
-   *
-   *     When the dialog is cancelable it can be closed by tapping the background or by pressing the back button on Android devices.
-   *   [/en]
+   *   [en]Whether the dialog is cancelable or not. A cancelable dialog can be closed by tapping the background or by pressing the back button on Android devices.[/en]
    *   [ja][/ja]
    */
   set cancelable(value) {
@@ -457,21 +453,21 @@ class AlertDialogElement extends BaseElement {
    * @type {Boolean}
    * @description
    *   [en]Whether the dialog is visible or not.[/en]
-   *   [ja]ダイアログが表示されているかどうか。[/ja]
+   *   [ja]要素が見える場合に`true`。[/ja]
    */
   get visible() {
     return this._visible;
   }
 
   /**
-   * @property backButtonHandler
+   * @property onDeviceBackButton
    * @readonly
    * @type {Object}
    * @description
    *   [en]Retrieve the back-button handler.[/en]
    *   [ja]バックボタンハンドラを取得します。[/ja]
    */
-  get backButtonHandler() {
+  get onDeviceBackButton() {
     return this._backButtonHandler;
   }