+++ core/src/elements/ons-modal/index.js
@@ -104,17 +104,17 @@ class ModalElement extends BaseElement {
 
 
   /**
-   * @property backButtonHandler
+   * @property onDeviceBackButton
    * @type {Object}
    * @description
    *   [en]Back-button handler.[/en]
    *   [ja]バックボタンハンドラ。[/ja]
    */
-  get backButtonHandler() {
+  get onDeviceBackButton() {
     return this._backButtonHandler;
   }
 
-  set backButtonHandler(handler) {
+  set onDeviceBackButton(handler) {
     if (this._backButtonHandler) {
       this._backButtonHandler.destroy();
     }
@@ -186,8 +186,8 @@ class ModalElement extends BaseElement {
    * @readonly
    * @type {Boolean}
    * @description
-   *   [en]Whether the dialog is visible or not.[/en]
-   *   [ja]ダイアログが表示されているかどうか。[/ja]
+   *   [en]Whether the element is visible or not.[/en]
+   *   [ja]要素が見える場合に`true`。[/ja]
    */
   get visible() {
     return this.style.display !== 'none';